+++ app/components/url/url-controller.js
@@ -12,42 +12,54 @@
  * state leads the url.
  */
 angular.module('lizard-nxt')
-  .controller('UrlController', ['$scope', 'LocationGetterSetter',
-  'UrlState', 'dataBounds', 'DataService', 'MapService', 'State', '$rootScope',
-  function ($scope, LocationGetterSetter, UrlState, dataBounds, DataService, MapService, State, $rootScope) {
+.controller('UrlController', [
+  '$scope',
+  'LocationGetterSetter',
+  'UrlState',
+  'dataBounds',
+  'DataService',
+  'MapService',
+  'State',
+  '$rootScope',
+  'LeafletService',
+  function (
+    $scope,
+    LocationGetterSetter,
+    UrlState,
+    dataBounds,
+    DataService,
+    MapService,
+    State,
+    $rootScope,
+    LeafletService
+  ) {
 
     // Configuration object for url state.
     var state = {
       context: { // Locally used name for the state
-        value: 'map', // Default to accomodate future support for contextSwitch
+        value: 'map', // default
         part: 'path', // Part of the url where this state is stored,
         index: 0, // Position of the state in the part
-        update: true // When false, $locationChangeSucces is cancelled
       },
       layerGroups: {
         part: 'path',
         index: 1,
-        update: true
       },
       boxType: {
         part: 'path',
         index: 2,
-        update: true
       },
       geom: {
         part: 'path',
         index: 3,
-        update: true
       },
       mapView: {
         part: 'at',
         index: 0,
-        update: true
       },
       timeState: {
         part: 'at',
         index: 1,
-        update: true
       }
     };
 
@@ -55,10 +67,10 @@ angular.module('lizard-nxt')
     * @function
     * @memberOf app.UrlController
     * @summary Enables or disables layerGroups on the basis of the url.
-    * @description Takes the layerGroups as defined in the url to turn layerGroups on
-    *              afterwards it initializes all other layerGroups. This is done
-    *              here so MapService does not turn on layerGroups which are
-    *              turned of later by this controller.
+    * @description Takes the layerGroups as defined in the url to turn
+    *              layerGroups on afterwards it initializes all other
+    *              layerGroups. This is done here so MapService does not turn
+    *              on layerGroups which are turned of later by this controller.
     * @param {string} String representation of layerGroups on url
     */
     var enablelayerGroups = function (layerGroupString) {
@@ -81,20 +93,29 @@ angular.module('lizard-nxt')
     * @param {string} String representation of mapView on url
     */
     var enableMapView = function (mapView) {
+      var map = LeafletService.map('url-temp-map');
       var fn = function () {
-        MapService.fitBounds(dataBounds);
+        map.fitBounds(LeafletService.latLngBounds(
+          L.latLng(dataBounds.south, dataBounds.east),
+          L.latLng(dataBounds.north, dataBounds.west)
+        ));
       };
 
       if (mapView) {
         var view = UrlState.parseMapView(mapView);
         if (view) {
-          MapService.setView(view.latLng, view.zoom, view.options);
+          map.setView(view.latLng, view.zoom, {});
         } else {
           fn();
         }
       } else {
         fn();
       }
+      State.spatial.bounds =  map.getBounds();
+      State.spatial.zoom =  map.getZoom();
+      map.remove();
+      var tempMap = document.getElementById('url-temp-map');
+      tempMap.parentNode.removeChild(tempMap);
     };
 
     /**
@@ -103,7 +124,6 @@ angular.module('lizard-nxt')
     $scope.$watch(State.toString('layerGroups.active'),
       function (n, o) {
         if (n === o) { return true; }
-        state.layerGroups.update = false;
         UrlState.setlayerGroupsUrl(state, State.layerGroups.active);
       }
     );
@@ -113,13 +133,12 @@ angular.module('lizard-nxt')
      */
     $scope.$watch(State.toString('spatial.bounds'), function (n, o) {
       if (n === o) { return true; }
-      state.mapView.update = false;
       UrlState.setCoordinatesUrl(state,
         State.spatial.bounds.getCenter().lat,
         State.spatial.bounds.getCenter().lng,
         State.spatial.zoom
       );
-    }, true);
+    });
 
     /**
      * Set timeState when timeState changed.
@@ -127,8 +146,11 @@ angular.module('lizard-nxt')
     $scope.$watch(State.toString('temporal.timelineMoving'), function (n, o) {
       if (n === o) { return true; }
       if (!State.temporal.timelineMoving) {
-        state.timeState.update = false;
-        UrlState.setTimeStateUrl(state, State.temporal.start, State.temporal.end);
+        UrlState.setTimeStateUrl(
+          state,
+          State.temporal.start,
+          State.temporal.end
+        );
       }
     });
 
@@ -137,7 +159,6 @@ angular.module('lizard-nxt')
      */
     $scope.$watch(State.toString('box.type'), function (n, old) {
       if (n === old) { return true; }
-      state.boxType.update = false;
       LocationGetterSetter.setUrlValue(
         state.boxType.part, state.boxType.index, State.box.type
       );
@@ -145,8 +166,20 @@ angular.module('lizard-nxt')
       if (old === 'point' || old === 'line') {
         // Remove geometry from url
         state.boxType.update = false;
-        LocationGetterSetter.setUrlValue(state.geom.part, state.geom.index, undefined);
+        LocationGetterSetter.setUrlValue(
+          state.geom.part, state.geom.index, undefined);
       }
+    });
+
+    /*
+     * Set context when context changed
+     */
+    $scope.$watch(State.toString('context'), function (n, old) {
+      if (n === old) { return true; }
+      state.context.update = false;
+      LocationGetterSetter.setUrlValue(
+        state.context.part, state.context.index, $scope.context
+      );
     });
 
     /**
@@ -155,7 +188,10 @@ angular.module('lizard-nxt')
     $scope.$watch(State.toString('spatial.here'), function (n, o) {
       if (n === o || State.box.type !== 'point') { return true; }
       state.geom.update = false;
-      UrlState.setgeomUrl(state, State.box.type, State.spatial.here, State.spatial.points);
+      UrlState.setgeomUrl(state,
+                          State.box.type,
+                          State.spatial.here,
+                          State.spatial.points);
     });
 
     /**
@@ -163,19 +199,18 @@ angular.module('lizard-nxt')
      */
     $scope.$watch(State.toString('spatial.points'), function (n, o) {
       if (n === o || State.box.type !== 'line') { return true; }
-      state.geom.update = false;
-      UrlState.setgeomUrl(state, State.box.type, State.spatial.here, State.spatial.points);
-    }, true);
+      UrlState.setgeomUrl(state,
+        State.box.type,
+        State.spatial.here,
+        State.spatial.points
+      );
+    });
 
     /**
      * Listener to update map view when user changes url
      *
      * $locationChangeSucces is broadcasted by angular
      * when the hashSyncHelper in util-service changes the url
-     *
-     * updateUrl is set to false when the application updates
-     * the url. Then, this listener is fired but does nothing but
-     * resetting the updateUrl back to true
      */
 
     // $locationChangeSuccess is fired once when this controller is initialized.
@@ -184,38 +219,40 @@ angular.module('lizard-nxt')
     // the app to update to the time of the url.
     State.temporal.timelineMoving = true;
 
-    $scope.$on('$locationChangeSuccess', function (e, oldurl, newurl) {
-      if (UrlState.update(state)) {
-        var boxType = LocationGetterSetter.getUrlValue(state.boxType.part, state.boxType.index),
-          geom = LocationGetterSetter.getUrlValue(state.geom.part, state.geom.index),
-          layerGroupsFromURL = LocationGetterSetter.getUrlValue(state.layerGroups.part, state.layerGroups.index),
-          mapView = LocationGetterSetter.getUrlValue(state.mapView.part, state.mapView.index),
-          time = LocationGetterSetter.getUrlValue(state.timeState.part, state.timeState.index),
-          context = LocationGetterSetter.getUrlValue(state.context.part, state.context.index);
+    var listener = $scope.$on('$locationChangeSuccess', function (e, oldurl, newurl) {
+      var boxType = LocationGetterSetter.getUrlValue(state.boxType.part, state.boxType.index),
+        geom = LocationGetterSetter.getUrlValue(state.geom.part, state.geom.index),
+        layerGroupsFromURL = LocationGetterSetter.getUrlValue(state.layerGroups.part, state.layerGroups.index),
+        mapView = LocationGetterSetter.getUrlValue(state.mapView.part, state.mapView.index),
+        time = LocationGetterSetter.getUrlValue(state.timeState.part, state.timeState.index),
+        context = LocationGetterSetter.getUrlValue(state.context.part, state.context.index);
 
+      if (context) {
+        State.context = context;
+      } else {
         LocationGetterSetter.setUrlValue(state.context.part, state.context.index, state.context.value);
-        if (boxType) {
-          State.box.type = boxType;
-        } else {
-          LocationGetterSetter.setUrlValue(state.boxType.part, state.boxType.index, State.box.type);
-        }
-        if (geom) {
-          State.spatial = UrlState.parseGeom(State.box.type, geom, State.spatial);
-        }
-        enablelayerGroups(layerGroupsFromURL);
-        enableMapView(mapView);
+      }
+      if (boxType) {
+        State.box.type = boxType;
+      } else {
+        LocationGetterSetter.setUrlValue(state.boxType.part, state.boxType.index, State.box.type);
+      }
+      if (geom) {
+        State.spatial = UrlState.parseGeom(State.box.type, geom, State.spatial);
+      }
+      enablelayerGroups(layerGroupsFromURL);
+      enableMapView(mapView);
 
-        if (time) {
-          State.temporal = UrlState.parseTimeState(time, State.temporal);
-        } else {
-          state.timeState.update = false;
-          UrlState.setTimeStateUrl(state, State.temporal.start, State.temporal.end);
-        }
+      if (time) {
+        State.temporal = UrlState.parseTimeState(time, State.temporal);
+      } else {
+        state.timeState.update = false;
+        UrlState.setTimeStateUrl(state, State.temporal.start, State.temporal.end);
       }
       State.temporal.timelineMoving = false;
-      angular.forEach(state, function (value) {
-        value.update = true;
-      });
+
+      listener(); // remove this listener
     });
+
   }
 ]);