+++ test/karma.conf.dev.js
@@ -34,7 +34,6 @@ module.exports = function(config) {
       'vendor/ng-csv/build/ng-csv.min.js',
       'vendor/leaflet.markercluster/dist/leaflet.markercluster.js',
       // endbower
-
       // application
       'test/mocks/beforeModuleMocks.js',
       'app/components/state/state.js', // Load these first to prevent dep clash
@@ -43,6 +42,8 @@ module.exports = function(config) {
       'app/components/state/state.js', // Load these first to prevent dep clash
       'app/templates.js',
       'app/components/omnibox/omnibox.js',
+      'app/components/state/state.js',
+      'app/components/dashboard/dashboard.js',
       'app/lizard-nxt.js',
       'test/mocks/**/*.js',
       'app/lib/**/*.js',