+++ CHANGES.rst
@@ -1,13 +1,35 @@
 Changelog of lizard-nxt client
 ==============================
 
-Unreleased (1.2.dev) (XXXX-XX-XX)
----------------------
+Unreleased (1.1.4) (XXXX-XX-XX)
+-------------------------------
 -
 
 
-Release 1.1.3 (2014-12-30)
+Release 1.2.1 (2015-1-8)
 ---------------------
+
+- Zapped patches for display_name vs name.
+
+- Beta dashboard implementation for events.
+
+
+Release 1.1.6 (2015-1-7)
+------------------------
+
+- Fix 'TODO' label for source in discrete raster point click.
+
+
+Release 1.1.5 (2015-1-7)
+------------------------
+
+- Fix category 'Overig' in hori. stacked bar charts
+
+- Fix client side handling of discrete rasters.
+
+
+Release 1.1.3 (2014-12-30)
+--------------------------
 
 - Fix 'Cannot read property 'lng' of undefined'.
 
@@ -21,9 +43,8 @@ Release 1.0.2 (2014-12-16)
 
 - Help button.
 
-
 Release 1.1.2 (2014-12-24)
----------------------
+--------------------------
 
 - handle API response for discrete rasters (same format for point/area mode)
 
@@ -57,6 +78,7 @@ Release 1.1.1 (2014-12-23)
 
 - Removed unused images.
 
+- Raster animation on day images.
 
 Release 1.1.0 (2014-12-17)
 --------------------------